Move i_listlen in struct inclist for better packing in 64-bit builds
Putting it with another int instead of between two pointers gets rid
of two places where 4 bytes each had to be padded into the structure
for alignment.

Reduces the size of an individual struct from 64 bytes to 56 bytes,
and of the inclist[ MAXFILES ] array from 128k to 112k.

Found by clang:
./def.h:111:18: warning: padding struct 'struct inclist' with 4 bytes to align 'i_defs' [-Wpadded]
        struct symtab   **i_defs;       /* symbol table for this file and its
                          ^
./def.h:114:12: warning: padding struct 'struct inclist' with 4 bytes to align 'i_merged' [-Wpadded]
        boolean         *i_merged;      /* whether we have merged child

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>